flaskアプリケーションに合わせてフロントエンドのディレクトリ構成を変更
