--- style.css
@@ -1,31 +1,0 @@
-/* style.css */
-body {
-    font-family: Arial, sans-serif;
-}
-
-#urlForm {
-    margin: 20px;
-    padding: 20px;
-    border: 1px solid #ddd;
-    display: flex;
-    flex-direction: column;
-    width: 300px;
-}
-
-#urlInput {
-    margin-bottom: 10px;
-    padding: 8px;
-    border: 1px solid #ccc;
-}
-
-#submitButton {
-    padding: 10px 20px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    cursor: pointer;
-}
-
-#submitButton:hover {
-    background-color: #0056b3;
-}
